backend/requirements.txt
# FastAPI フレームワーク
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# データベース
aiomysql==0.2.0
pymysql>=1.1.0
SQLAlchemy>=2.0.25
alembic==1.13.1
Mako>=1.2.0

# 認証・セキュリティ
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=3.2.0,<4.0.0  # passlib 1.7.4 と完全互換性のあるバージョン
python-dotenv==1.0.0

# データバリデーション
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator==2.1.0

# 日時処理
python-dateutil==2.8.2
pytz==2024.1

# CORS
fastapi-cors==0.0.6

# ロギング
loguru==0.7.2

# テスト
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# 文件监视 (BT-data 受信 CSV 同步 / 生産計画 Excel 同步)
watchdog>=3.0.0
mysql-connector-python>=8.0.0
openpyxl>=3.1.0

# コード品質
black==24.1.1
flake8==7.0.0
mypy==1.8.0

